refactor : rename utils package to util
